+++ src/main/java/com/example/auction/Global/error/handler/GlobalExceptionHandler.java
@@ -1,11 +1,8 @@
 package com.example.auction.Global.error.handler;
 
-import com.example.auction.Global.CommonResponseBody;
 import com.example.auction.Global.error.exception.CustomException;
 import com.example.auction.Global.error.response.ErrorResponse;
 import com.example.auction.Global.error.response.ErrorResponseDto;
-import io.jsonwebtoken.ExpiredJwtException;
-import io.jsonwebtoken.JwtException;
 import io.swagger.v3.oas.annotations.Hidden;
 import jakarta.validation.ConstraintViolationException;
 import lombok.extern.slf4j.Slf4j;
@@ -14,9 +11,6 @@ import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.HttpStatusCode;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.authentication.BadCredentialsException;
-import org.springframework.security.authorization.AuthorizationDeniedException;
-import org.springframework.security.core.AuthenticationException;
 import org.springframework.validation.FieldError;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
@@ -25,7 +19,6 @@ import org.springframework.web.bind.annotation.RestControllerAdvice;
 import org.springframework.web.context.request.WebRequest;
 import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
 
-import java.nio.file.AccessDeniedException;
 import java.time.LocalDateTime;
 
 import static com.example.auction.Global.error.errorcode.ErrorCode.DUPLICATE_RESOURCE;
@@ -62,50 +55,5 @@ public class GlobalExceptionHandler extends ResponseEntityExceptionHandler {
             errorResponseDto.addValidationError(fieldError.getField(), fieldError.getDefaultMessage());
         }
         return ResponseEntity.unprocessableEntity().body(errorResponseDto);
-    }
-
-    // Security와 관련된 AuthenticationException 예외 처리.
-    @ExceptionHandler(AuthenticationException.class)
-    protected ResponseEntity<CommonResponseBody<Void>> handleAuthException(
-            AuthenticationException e) {
-        HttpStatus statusCode = e instanceof BadCredentialsException
-                ? HttpStatus.FORBIDDEN
-                : HttpStatus.UNAUTHORIZED;
-
-        return ResponseEntity
-                .status(statusCode)
-                .body(new CommonResponseBody<>(e.getMessage()));
-    }
-
-    // Security와 관련된 AccessDeniedException 예외 처리.
-    @ExceptionHandler(AccessDeniedException.class)
-    protected ResponseEntity<CommonResponseBody<Void>> handleAccessDeniedException(
-            AccessDeniedException e) {
-
-        return ResponseEntity
-                .status(HttpStatus.UNAUTHORIZED)
-                .body(new CommonResponseBody<>(e.getMessage()));
-    }
-
-    // Security와 관련된 AuthorizationDeniedException 예외 처리.
-    @ExceptionHandler(AuthorizationDeniedException.class)
-    protected ResponseEntity<CommonResponseBody<Void>> handleAuthorizationDeniedException(
-            AuthorizationDeniedException e) {
-
-        return ResponseEntity
-                .status(HttpStatus.UNAUTHORIZED)
-                .body(new CommonResponseBody<>(e.getMessage()));
-    }
-
-    // JWT와 관련된 JwtException 예외 처리.
-    @ExceptionHandler(JwtException.class)
-    protected ResponseEntity<CommonResponseBody<Void>> handleJwtException(JwtException e) {
-        HttpStatus httpStatus = e instanceof ExpiredJwtException
-                ? HttpStatus.UNAUTHORIZED
-                : HttpStatus.FORBIDDEN;
-
-        return ResponseEntity
-                .status(httpStatus)
-                .body(new CommonResponseBody<>(e.getMessage()));
     }
 }
