+++ src/main/java/com/example/auction/Product/ProductService.java
@@ -2,6 +2,7 @@ package com.example.auction.Product;
 
 import com.example.auction.Product.Dto.ProductRequestDto;
 import com.example.auction.Product.Dto.ProductResponseDto;
+import com.example.auction.User.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
