+++ src/main/java/com/example/auction/User/User.java
@@ -1,0 +1,27 @@
+package com.example.auction.User;
+
+import com.example.auction.Global.BaseEntity;
+import jakarta.persistence.*;
+import lombok.Getter;
+
+@Entity
+@Getter
+@Table(name = "user")
+public class User extends BaseEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private String password;
+
+    private String name;
+
+    private String phoneNumber;
+
+    private Long point;
+
+    private String roll;
+
+    private String status;
+}
