+++ src/main/java/com/example/auction/User/UserRepository.java
@@ -1,0 +1,4 @@
+package com.example.auction.User;
+
+public interface UserRepository {
+}
