+++ src/main/java/com/example/auction/User/UserService.java
@@ -1,0 +1,4 @@
+package com.example.auction.User;
+
+public class UserService {
+}
