--- src/main/java/com/example/auction/User/controller/UserController.java
@@ -1,35 +1,0 @@
-package com.example.auction.User.controller;
-
-import com.example.auction.Global.CommonResponseBody;
-import com.example.auction.User.dto.UserResponseDto;
-import com.example.auction.User.service.UserService;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@Slf4j(topic = "Security::UserController")
-@RestController
-@RequestMapping(value = "/user")
-@RequiredArgsConstructor
-public class UserController {
-
-    private final UserService userService;
-
-    /**
-     * USER 권한으로 접근 테스트.
-     *
-     * @return {@code ResponseEntity<CommonResponseBody<MemberResponse>>}
-     */
-    @GetMapping("/something")
-    public ResponseEntity<CommonResponseBody<UserResponseDto>> doSomethingAsUser() {
-        UserResponseDto dto = this.userService.doSomethingAsUser();
-
-        return ResponseEntity
-                .status(HttpStatus.OK)
-                .body(new CommonResponseBody<>("USER 권한으로 접근 성공.", dto));
-    }
-}
