--- src/main/java/com/example/auction/User/dto/AccountRequestDto.java
@@ -1,42 +1,0 @@
-package com.example.auction.User.dto;
-
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotBlank;
-import lombok.Getter;
-
-/**
-
- * <p> 로그인 request를 위한 DTO. </p>
- *
- * @author Seokgyu Hwang (Chris)
- * @version 1.0
- * @since 1.0
- */
-@Getter
-public class AccountRequestDto {
-
-  /**
-   * 이메일 (Spring Security에서 username으로 식별하기 위함).
-   */
-  @NotBlank(message = "email은 빈 값이 허용되지 않습니다.")
-  @Email(message = "올바른 email 형식이 아닙니다.")
-  private final String email;
-
-  /**
-   * 암호.
-   */
-  @NotBlank(message = "password는 빈 값이 허용되지 않습니다.")
-  private final String password;
-
-  /**
-   * 생성자.
-   *
-   * @param email    이메일
-   * @param password 암호
-   */
-  public AccountRequestDto(String email, String password) {
-    this.email = email;
-    this.password = password;
-
-  }
-}
