--- src/main/java/com/example/auction/User/entity/Role.java
@@ -1,33 +1,0 @@
-package com.example.auction.User.entity;
-
-import lombok.Getter;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import java.util.List;
-
-@Getter
-public enum Role {
-
-    USER("user"),
-
-    ADMIN("admin");
-
-    private final String name;
-
-    Role(String name) {
-        this.name = name;
-    }
-
-    public static Role of(String roleName) throws IllegalArgumentException {
-        for (Role role : values()) {
-            if (role.getName().equals(roleName.toLowerCase())) {
-                return role;
-            }
-        }
-
-        throw new IllegalArgumentException("해당하는 이름의 권한을 찾을 수 없습니다: " + roleName);
-    }
-
-    public List<SimpleGrantedAuthority> getAuthorities() {
-        return List.of(new SimpleGrantedAuthority("ROLE_" + this.name()));
-    }
-}
