+++ src/main/resources/application.properties
@@ -3,24 +3,18 @@ spring.application.name=auction
 spring.config.import=optional:file:.env[.properties]
 
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
-spring.datasource.url=jdbc:mysql://localhost:3306/auction?serverTimezone=UTC&characterEncoding=UTF-8
-spring.datasource.username=root
-spring.datasource.password=wlstjr0114
+spring.datasource.url=${MYSQL_URL}
+spring.datasource.username=${MYSQL_USER}
+spring.datasource.password=${MYSQL_PASSWORD}
 
 spring.jpa.show-sql=true
 spring.jpa.database-platform=org.hibernate.dialect.MySQL8Dialect
 
 # create, update, none, creat-drop
 spring.jpa.database=mysql
-spring.jpa.hibernate.ddl-auto=create
+spring.jpa.hibernate.ddl-auto=${JPA_HIBERNATE_DDL}
 spring.jpa.hibernate.naming.physical-strategy=org.hibernate.boot.model.naming.CamelCaseToUnderscoresNamingStrategy
 spring.jpa.generate-ddl=false
 spring.jpa.properties.hibernate.format_sql=true
 spring.jpa.properties.hibernate.enable_lazy_load_no_trans=true
 spring.jpa.properties.hibernate.use_sql_comments=true
-
-jwt.secret=d4e8002d9a40324a4bc163505c22a1a8e0e77eb53d59efbd4f6e5a91ceb17736
-jwt.expiry-millis=300000
-
-logging.level.com.auction : DEBUG
-logging.level.org.springframework.security: DEBUG
